+++ provider/provider.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"strings"
 
-	"github.com/alexfalkowski/konfig/server/config/provider/env"
-	"github.com/alexfalkowski/konfig/server/config/provider/ssm"
-	"github.com/alexfalkowski/konfig/server/config/provider/vault"
+	"github.com/alexfalkowski/konfig/provider/env"
+	"github.com/alexfalkowski/konfig/provider/ssm"
+	"github.com/alexfalkowski/konfig/provider/vault"
 )
 
 const argumentsLen = 2
