+++ provider/vault/client.go
