+++ provider/vault/trace/opentracing/opentracing.go
