--- server/config/errors/errors.go
@@ -1,16 +1,0 @@
-package errors
-
-import (
-	"errors"
-)
-
-var (
-	// ErrUnmarshalError in config.
-	ErrUnmarshalError = errors.New("unmarshal issue")
-
-	// ErrMarshalError in config.
-	ErrMarshalError = errors.New("marshal issue")
-
-	// ErrTraverseError in config.
-	ErrTraverseError = errors.New("traverse issue")
-)
